minor change to update interface to use long as param type for durationmillsecs

+++ Joyride/Platforms/Android/AndroidScreen.cs
@@ -152,7 +152,7 @@ namespace Joyride.Platforms.Android
             return (value == "true");
         }
 
-        public override Screen Pull(Direction direction, int durationMillSecs = 1000)
+        public override Screen Pull(Direction direction, long durationMillSecs = 1000)
         {
             var screen = base.Pull(direction, durationMillSecs);
             Driver.WaitFor(TimeSpan.FromMilliseconds(500));
